+++ examples/react/src/main.tsx
@@ -4,8 +4,6 @@ import { CounterButton } from './comps/counter-button';
 import { Description } from './comps/description';
 import './main.css';
 
-const Clock = React.lazy(() => import('./comps/clock'));
-
 export function Main() {
-  return <Suspense fallback={'loading...'}><CounterButton /><Description /><div><Clock /></div></Suspense>
+  return <><CounterButton /><Description /><div></div></>
 }